refactoring

+++ Application/Consumptions/ConsumptionService.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 using Money2.Domain.Consumptions;
 
 namespace Money2.Application.Consumptions
@@ -16,7 +17,7 @@ namespace Money2.Application.Consumptions
         public void DeleteConsumption( Int32 userId, Int32 consumptionId )
         {
             var consumption = _consumptionRepository.Get( consumptionId );
-            if ( consumption != null || consumption.UserId == userId )
+            if ( consumption != null && consumption.UserId == userId )
             {
                 _consumptionRepository.Delete( consumption );
             }
@@ -24,7 +25,22 @@ namespace Money2.Application.Consumptions
 
         public List<ConsumptionDto> GetConsumptions( Int32 userId, DateTime? startDate, DateTime? endDate, Int32 typeId )
         {
-            throw new NotImplementedException();
+            return _consumptionRepository.GetConsumptions( userId, startDate, endDate )
+                .ConvertAll( c => new ConsumptionDto()
+                {
+                    Id = c.Id,
+                    Date = c.Date,
+                    Place = c.Place,
+                    Sum = c.Sum,
+                    UserId = c.UserId,
+                    ConsumptionItems = c.ConsumptionItems.ConvertAll( ci => new ConsumptionItemDto()
+                    {
+                        Id = ci.Id,
+                        Name = ci.Name,
+                        Price = ci.Price,
+                        Quantity = ci.Quantity
+                    } )
+                } );
         }
 
         public ConsumptionDto GetConsumption( Int32 userId, Int32 consumptionId )
@@ -43,7 +59,14 @@ namespace Money2.Application.Consumptions
                 Date = consumption.Date,
                 Place = consumption.Place,
                 Sum = consumption.Sum,
-                UserId = consumption.UserId
+                UserId = consumption.UserId,
+                ConsumptionItems = consumption.ConsumptionItems.ConvertAll( ci => new ConsumptionItemDto()
+                {
+                    Id = ci.Id,
+                    Name = ci.Name,
+                    Price = ci.Price,
+                    Quantity = ci.Quantity
+                } )
             };
         }
 
@@ -64,12 +87,26 @@ namespace Money2.Application.Consumptions
                 Id = dto.Id,
                 Date = dto.Date,
                 Sum = dto.Sum,
-                UserId = dto.UserId
+                Place = dto.Place,
+                UserId = dto.UserId,
+                ConsumptionItems = dto.ConsumptionItems.ConvertAll( ci => new ConsumptionItem()
+                {
+                    Name = ci.Name,
+                    Price = ci.Price,
+                    Quantity = ci.Quantity
+                } )
             };
 
-            //consumption.Sum = consumption.ConsumptionItems.Sum( x => x.Quantity * x.Price );
+            consumption.Sum = dto.ConsumptionItems.Sum( x => x.Quantity * x.Price );
+
+            if( consumption.Id != 0 )
+            {
+                _consumptionRepository.Clear( consumption.Id );
+            }
 
             _consumptionRepository.Save( consumption );
+
+            dto.Id = consumption.Id;
         }       
     }
 }
